+++ server/routes/documents.js
@@ -6,6 +6,7 @@ const path = require("path");
 
 const rootPath = path.join(__dirname, "../../");
 const dataFolderPath = path.join(rootPath, "data");
+const autosavesFolderPath = path.join(dataFolderPath, "autosaves");
 const notebooksFolderPath = path.join(dataFolderPath, "notebooks");
 const imageFolderPath = path.join(dataFolderPath, "images");
 const attachmentsFolderPath = path.join(dataFolderPath, "attachments");
@@ -374,38 +375,6 @@ router.put("/api/documents/updateFile", async (req, res) => {
 
   fs.writeFileSync(filePath, JSON.stringify(fileData, null, 2), "utf8");
   res.json({ success: true });
-});
-
-// Autosaves
-router.post("/api/documents/autosave", async (req, res) => {
-  const { name, folderPath, saveData } = req.body;
-  const defaultContent = JSON.stringify(
-    [
-      {
-        title: name,
-        content: saveData,
-      },
-    ],
-    null,
-    2,
-  );
-
-  try {
-    if (name && folderPath) {
-      const location = path.join(notebooksFolderPath, folderPath, name);
-      fs.writeFileSync(location, defaultContent, "utf8");
-      res.json({ success: true });
-    } else if (name) {
-      console.error("No Folder Path found.");
-    } else if (folderPath) {
-      console.error("No Name found.");
-    } else {
-      console.error("Required values not found for operation.");
-    }
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "Failed to create Autosave" });
-  }
 });
 
 router.post(